Too large, cumulative update
It started out as a rank fix in Holding.  This causes lots of changes in LoopHold cases.
This uncovered some other bugs.  In the end there is one case left in Holding that is
wrong (0x1445, 13 / 0x3a9417).  The reason is in MergeDefender to begin with, and
there is an alternative Compare function that corrects the problem.  But it causes other
problems.  An attempt to fix this problem is the new CompareInit() function in TrickList.
This has not been tested yet on all suits.  Instead there is a wrong entry for 0x1445.
But this was getting out of control, so I check it in now.

12,312 moves
0.4905, 4.7154
4/6
23,887 left in G51 to spell out.