Introduced a new rank adjustment in DefList
If a defense that is never chosen nonetheless uses a lower rank,
and #tricks is the same and alt length is 1, then adjust down.
Not sure if this is the right place for it in the long term, but it is
a modest improvement.  Led to a small number of simplifications
in LoopHold.
4,723,210 lines, 157,000 cases.